Massive Reorginization to Support RxJS Video Controls
All components and code related to the video playback have been abstracted into a feature module, and wired into a dedicated ngRx feature.

Application now features play/pause controls that dispatch actions, which modify the player state, and cause the player directive to pause the rendered HTMLVideoElement.

Application has been organized to be more like something Todd Moto would write, and as such it's easier to find what you're looking for.
