--- src/app/shared/modules/materialdesigncomponentsmodule/material-design-components.module.ts
@@ -1,25 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-// Import Material Components
-import { MatToolbarModule, MatCardModule, MatDividerModule, MatListModule, MatButtonModule } from '@angular/material';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    MatToolbarModule,
-    MatCardModule,
-    MatDividerModule,
-    MatListModule,
-    MatButtonModule
-  ],
-  exports: [
-    MatToolbarModule,
-    MatCardModule,
-    MatDividerModule,
-    MatListModule,
-    MatButtonModule
-  ],
-  declarations: []
-})
-export class MaterialDesignComponentsModule { }
