--- src/app/shared/player-model.ts
@@ -1,4 +1,0 @@
-export enum PlayerStatus {
-  PLAYING,
-  PAUSED
-}
