--- src/app/shared/store/models/stream.ts
@@ -1,15 +1,0 @@
-export enum StreamStatus {
-  LOADING,
-  ERROR,
-  PLAYING,
-  PAUSED,
-  SEEKING
-}
-
-export interface Stream {
-  src:   string,
-  title: string,
-  levels?: any,
-  playbackLevel?: any,
-  status?: StreamStatus
-}
