--- src/app/shared/store/models/streams.ts
@@ -1,4 +1,0 @@
-import { Stream } from './stream';
-export interface Streams {
-  streams: Stream[]
-}
