--- src/app/shared/store/models/video-player.ts
@@ -1,14 +1,0 @@
-export enum VideoPlayerStatus {
-  LOADING,
-  ERROR,
-  PLAYING,
-  PAUSED,
-  SEEKING
-}
-
-export interface VideoPlayer {
-  status: VideoPlayerStatus,
-  duration?: number,
-  selectedLevel?: any,
-  levels?: any
-}
