--- src/app/shared/store/reducers/selected-stream.reducers.ts
@@ -1,13 +1,0 @@
-import * as SelectedStreamActions from '../actions/selected-stream.actions';
-import { Stream } from '../models/stream';
-
-
-export const selectedStreamReducer = (state: Stream, action: SelectedStreamActions.Actions) => {
-  switch (action.type) {
-    case SelectedStreamActions.SET_SELECTED_STREAM:
-      return { ...state, ...action.payload };
-    default:
-      return state;
-  }
-}
-
