--- src/app/shared/store/store.reducers.ts
@@ -1,9 +1,0 @@
-import { streamsReducer } from './reducers/streams.reducer';
-import { selectedStreamReducer } from './reducers/selected-stream.reducers';
-import { reducer as videoPlayerReducer } from './reducers/video-player.reducer';
-
-export const reducers = {
-  streams: streamsReducer,
-  selectedStream: selectedStreamReducer,
-  videoPlayer: videoPlayerReducer
-};
