--- src/app/stream-list/stream-list.component.html
@@ -1,12 +1,0 @@
-<div class="stream-list-container">
-  <mat-list>
-    <p class="mat-title">Available Streams: </p>
-      <mat-list-item *ngFor="let stream of streams$ | async">
-        <button mat-button (click)="setSelectedStream(stream)">
-          <i class="material-icons">play_arrow</i>
-          {{ stream.title }}
-        </button>
-      </mat-list-item>
-    <mat-divider></mat-divider>
-  </mat-list>
-</div>
