--- src/app/video-controls/video-controls.component.html
@@ -1,4 +1,0 @@
-<div class="video-controls-container">
-  <video-controls-play></video-controls-play>
-  <video-controls-pause></video-controls-pause>
-</div>
