--- src/app/video-controls/video-controls.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-video-controls',
-  templateUrl: './video-controls.component.html',
-  styleUrls: ['./video-controls.component.scss']
-})
-export class VideoControlsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
