--- src/app/video-player/video-player.component.html
@@ -1,10 +1,0 @@
-<div class="video-player-container" *ngIf="selectedStream$ | async">
-  <div class="content">
-    <div>
-      <video appVideoPlayer></video>
-    </div>
-    <h1 class="mat-title"> {{ selectedStream$.title }}</h1>
-  </div>
-  <app-video-controls></app-video-controls>
-  <mat-divider></mat-divider>
-</div>
