+++ src/video-player/components/stream-list/stream-list.component.scss
