+++ src/video-player/components/video-control-panel/video-controls.component.html
@@ -1,0 +1,4 @@
+<div class="video-controls-container">
+  <player-controls-play></player-controls-play>
+  <player-controls-pause></player-controls-pause>
+</div>
