+++ src/video-player/components/video-control-panel/video-controls.component.scss
