+++ src/video-player/components/video-control-panel/video-controls.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'video-control-panel',
+  templateUrl: './video-controls.component.html',
+  styleUrls: ['./video-controls.component.scss']
+})
+export class VideoControlsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
