+++ src/video-player/components/video-player/video-player.component.scss
