+++ src/video-player/store/effects/stream.effects.ts
@@ -1,0 +1,30 @@
+import { Injectable } from '@angular/core';
+import { Action } from '@ngrx/store';
+import { Actions, Effect, ofType } from '@ngrx/effects';
+import { Observable, of } from 'rxjs';
+import { map } from 'rxjs/operators';
+import { Stream } from '../models/stream';
+import * as StreamActions from '../actions/stream.actions';
+
+const streams: Stream[] = [
+    {
+      title: 'Sintel',
+      src: 'https://bitdash-a.akamaihd.net/content/sintel/hls/playlist.m3u8'
+    },
+    {
+      title: 'Big Buck Bunny',
+      src: 'https://video-dev.github.io/streams/x36xhzz/x36xhzz.m3u8'
+    }
+  ];
+
+
+@Injectable()
+export class StreamEffects {
+  @Effect()
+  fetch$: Observable<Action> = this.actions$.pipe(
+    ofType(StreamActions.FETCH_STREAMS),
+    map(action => ({ type: StreamActions.SET_STREAMS, payload: streams }))
+  );
+
+  constructor(private actions$: Actions) {}
+}
