+++ src/video-player/store/models/player.ts
@@ -1,0 +1,14 @@
+export enum PlayerStatus {
+  LOADING,
+  ERROR,
+  PLAYING,
+  PAUSED,
+  SEEKING
+}
+
+export interface Player {
+  status: PlayerStatus,
+  duration?: number,
+  selectedLevel?: any,
+  levels?: any
+}
