+++ src/video-player/store/models/stream.ts
@@ -1,0 +1,19 @@
+export enum StreamStatus {
+  LOADING,
+  ERROR,
+  PLAYING,
+  PAUSED,
+  SEEKING
+}
+
+export interface Stream {
+  src:   string,
+  title: string,
+  levels?: any,
+  playbackLevel?: any,
+  status?: StreamStatus
+}
+
+export interface StreamList {
+  streams: Stream[]
+}
