+++ src/video-player/store/store.index.ts
@@ -1,0 +1,6 @@
+import * as reducers from './reducers';
+
+export const store = {
+  reducers
+};
+
